requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
python-jose 
passlib[bcrypt]
python-multipart